Методы расширения, связанные с HTTP вынесены в Infrastructure
