+++ src/PhotoCatalog.Domain/Extensions/ResultExtensions.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 
-using Microsoft.AspNetCore.Http;
-
 using PhotoCatalog.Domain.Primitives;
 
 namespace PhotoCatalog.Domain.Extensions;
@@ -438,15 +436,5 @@ public static class ResultExtensions
         return result.IsSuccess
             ? success(result.Value!)
             : failure(result.Error);
-    }
-
-    /// <summary>
-    ///     Преобразует <see cref="Result{T}" /> в <see cref="IResult" /> с соответствующим HTTP-статусом.
-    /// </summary>
-    public static IResult ToHttpResult<T>(this Result<T> result)
-    {
-        return result.IsSuccess
-            ? result.Value is null ? Results.NoContent() : Results.Ok(result.Value)
-            : result.Error.ToHttpResult();
     }
 }
