+++ src/PhotoCatalog.Domain/Extensions/ResultVoidExtensions.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 
-using Microsoft.AspNetCore.Http;
-
 using PhotoCatalog.Domain.Primitives;
 
 namespace PhotoCatalog.Domain.Extensions;
@@ -226,15 +224,5 @@ public static class ResultVoidExtensions
         return result.IsSuccess
             ? success()
             : failure(result.Error);
-    }
-
-    /// <summary>
-    ///     Преобразует <see cref="ResultVoid" /> в <see cref="IResult" />.
-    /// </summary>
-    public static IResult ToHttpResult(this ResultVoid result)
-    {
-        return result.IsSuccess
-            ? Results.NoContent()
-            : result.Error.ToHttpResult();
     }
 }
