+++ src/PhotoCatalog.Domain/PhotoCatalog.Domain.csproj
@@ -14,8 +14,4 @@
     </Content>
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" />
-  </ItemGroup>
-
 </Project>
