+++ src/PhotoCatalog.Infrastructure/PhotoCatalog.Infrastructure.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <PackageReference Include="Dapper"/>
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" />
     <PackageReference Include="Microsoft.Data.Sqlite.Core"/>
     <PackageReference Include="Microsoft.Extensions.Logging"/>
     <PackageReference Include="Microsoft.Extensions.Caching.Hybrid"/>
